[Add-on] v3.0.0 - details in description
- Styles are now managed completely in javascript, which means there
will no longer be any external stylesheets.
- Removed notify() helper function.
- a globally exposed Notice constructor will be used to create new
notices.
- Styles/Markup/Defaults are now fully editable via properties exposed
on the Notice constructor object.
- Animation speed can now be controlled.
- a newly generated notice will be appended to the DOM and await the
its prompt() method to be invoked which will return a promise.
- The promise returned by prompt will not reject anymore but instead
will be resolved when an action is clicked with the promise’s value
resolving the action’s name.
- The Notice constructor will no longer accept a resolver/rejecter
argument due to the previously noted change.
- Passing highlight:true to the Notice constructor allows you to
specify which button should be highlighted.
- Passing keepAlive:true to the Notice constructor will prevent the
notice from beyond removed from the DOM after it is dismissed and will
instead be kept there for future re-prompts.
- Notices will now be resolved after their animation completes.